+++ pages/empty.js
@@ -1,9 +1,7 @@
 import EmptyMovieList from '../components/EmptyMoviesList/EmptyMoviesList';
-import MainHeader from '../components/MainHeader/MainHeader';
 
 const EmptyMoviesListPage = () => (
   <header className='header-content'>
-    <MainHeader />
     <EmptyMovieList />
   </header>
 );
