+++ pages/movies/[id].js
@@ -1,48 +1,40 @@
 import MovieDetails from '../../components/MovieDetails/MovieDetails';
 import MovieDetailsHeader from '../../components/MovieDetailsHeader/MovieDetailsHeader';
 import Home from '../../components/Home/Home';
-import { useRouter } from 'next/router';
 import { useEffect } from 'react';
-import { connect, useSelector } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import { wrapper } from '../../store';
-import { getMovies } from '../../actions';
+import { useDispatch } from 'react-redux';
 import Head from 'next/head';
-import { moviesSelector } from '../../selectors/moviesSelectors';
+import { fetchMovie, fetchMovies } from '../../actions';
+import { FETCH_DATA_SUCCESS } from '../../redux/types';
 
-const MoviePage = (props) => {
-  const {
-    query: { id },
-  } = useRouter();
-
-  const movies = useSelector(moviesSelector);
-  const { title } = movies.find((movie) => movie.id === id);
-
+const MoviePage = ({ movie, movies }) => {
+  const dispatch = useDispatch();
   useEffect(() => {
-    props.getMovies();
-  }, [id]);
-
+    dispatch({ type: FETCH_DATA_SUCCESS, payload: movies });
+  }, []);
   return (
     <>
       <Head>
-        <title>{title}</title>
+        <title>{movie.title}</title>
         <meta name='viewport' content='initial-scale=1.0, width=device-width' />
       </Head>
       <MovieDetailsHeader />
-      <MovieDetails id={id} />
+      <MovieDetails movie={movie} />
       <Home />
     </>
   );
 };
 
-export const getServerSideProps = wrapper.getServerSideProps(async ({ store }) => {
-  store.dispatch(getMovies());
-});
+export const getServerSideProps = async ({ query }) => {
+  const movies = await fetchMovies();
+  const movie = await fetchMovie(query.id);
 
-const mapDispatchToProps = (dispatch) => {
   return {
-    getMovies: bindActionCreators(getMovies, dispatch),
+    props: {
+      movie,
+      movies,
+    },
   };
 };
 
-export default connect(null, mapDispatchToProps)(MoviePage);
+export default MoviePage;
