+++ redux/reducers/modalsReducer.js
