+++ redux/selectors/filterSelectors.js
