+++ redux/selectors/sortingSelectors.js
