--- utils/mockedData.js
@@ -1,82 +1,0 @@
-// export const mockedData = [
-//   {
-//     title: "Green mile",
-//     genre: "Drama, Fantasy",
-//     releaseDate: 1994,
-//     id: "1",
-//     img: "https://pbs.twimg.com/media/ELeJbCYXkAE18WI.jpg",
-//   },
-//   {
-//     title: "Interstellar",
-//     genre: "Drama, Fantasy, Adventures",
-//     releaseDate: 2014,
-//     id: "2",
-//     img: "https://mir-s3-cdn-cf.behance.net/project_modules/1400_opt_1/b0f70f29979965.560d298403bc7.jpg",
-//   },
-//   {
-//     title: "Forrest Gump",
-//     genre: "Drama, Comedy",
-//     releaseDate: 1994,
-//     id: "3",
-//     img: "https://pbs.twimg.com/media/EK7Tf-IXkAErCgW.jpg",
-//   },
-//   {
-//     title: "Shutter Island",
-//     genre: "Drama, Thriller",
-//     releaseDate: 2009,
-//     id: "4",
-//     img: "https://pbs.twimg.com/media/EHb-eADX4AAP_6T.jpg",
-//   },
-//   {
-//     title: "The Matrix",
-//     genre: "Fantasy, Thriller",
-//     releaseDate: 1999,
-//     id: "5",
-//     img: "https://cinelog.net/wp-content/uploads/2020/04/aOIuZAjPaRIE6CMzbazvcHuHXDc-768x1152.jpg",
-//   },
-//   {
-//     title: "The brother",
-//     genre: "Drama, Thriller",
-//     releaseDate: 1997,
-//     id: "6",
-//     img: "https://i10.fotocdn.net/s124/a80214fc2d561294/user_xl/2825091898.jpg",
-//   },
-// ];
-
-export const mockedData = [
-  {
-    id: "-MWJmKyNgxg1U3WKo2MF",
-    description:
-      "The Green Mile is a 1999 American fantasy drama film writtenand directed by Frank Darabont and based on Stephen Kings 1996 novel of the same name. It stars Tom Hanks as a death row corrections officer during the Great Depression who witnesses supernatural events that occur after an enigmatic inmate (Michael Clarke Duncan) is brought to his facility. David Morse, Bonnie Hunt, Doug Hutchison and James Cromwell appear in supporting roles.The film premiered on December 10, 1999, in the United States to positive reviews from critics, who praised its visual style and performances. It was a commercial success, grossing $286 million from its $60 million budget, and was nominated for four Academy Awards: Best Picture, Best Supporting Actor for Duncan, Best Sound and Best Screenplay Based on Material Previously Produced or Published.",
-    duration: "120",
-    genre: "Drama, Fantasy",
-    img: "https://pbs.twimg.com/media/ELeJbCYXkAE18WI.jpg",
-    rating: "4.9",
-    releaseDate: "1994",
-    title: "Green mile",
-  },
-
-  {
-    id: "-MWJof12zk_84pRdWPsq",
-    description:
-      "Forrest Gump is a 1994 American romantic comedy-drama film directed by Robert Zemeckis and written by Eric Roth. It is based on the 1986 novel of the same name by Winston Groom and stars Tom Hanks, Robin Wright, Gary Sinise, Mykelti Williamson and Sally Field. The story depicts several decades in the life of Forrest Gump (Hanks), a slow-witted but kind-hearted man from Alabama who witnesses and unwittingly influences several defining historical events in the 20th century United States. The film differs substantially from the novel. Principal photography took place between August and December 1993, mainly in Georgia, North Carolina and South Carolina. Extensive visual effects were used to incorporate Hanks into archived footage and to develop other scenes. The soundtrack features songs reflecting the different periods seen in the film.",
-    duration: "125",
-    genre: "Comedy",
-    img: "https://pbs.twimg.com/media/EK7Tf-IXkAErCgW.jpg",
-    rating: "4.3",
-    releaseDate: "1994-12-25",
-    title: "Forrest Gump",
-  },
-
-  {
-    id: "-MWJn2JOvxRpVKrK0Ye3",
-    description:
-      "Brother (Russian: Брат, translit. Brat) is a 1997 Russian neo-noir crime drama film written and directed by Aleksei Balabanov. The film stars Sergei Bodrov Jr. as Danila Bagrov, a young ex-army conscript who becomes embroiled with the Saint Petersburg mob through his criminal older brother. It appeared in the Un Certain Regard section at the 1997 Cannes Film Festival.[1]After its release on VHS in June 1997, Brother unexpectedly became one of the most commercially successful Russian films of the 1990s and quickly rose to cult film status throughout Russia. Due to the films popularity and fan demand, a sequel, Brother 2, was released in 2000",
-    duration: "129",
-    genre: "Thriller",
-    img: "https://i10.fotocdn.net/s124/a80214fc2d561294/user_xl/2825091898.jpg",
-    rating: "4.5",
-    releaseDate: "1997-06-01",
-    title: "Brother",
-  },
-];
